21.1.Functional Testing with Emails
When we originally added our Mailtrap config... I was a bit lazy. I put the value into .env. But because that file is committed... we really shouldn't put any sensitive values into it. Well, you could argue that Mailtrap credentials aren't that sensitive, but let's fix this properly. Copy the MAILER_DSN and open .env.local.
If you don't have a .env.local file yet, just create it. I already have one so that I can customize my local database config. The values in this file override the ones in .env. And because this file is ignored by .gitignore, these values won't be committed.

Back in .env, let's set MAILER_DSN back to the original value, which was smtp://localhost.
And yes, this does mean that when a developer clones the project, unless they customize MAILER_DSN in their own .env.local file, they'll get an error if they try to register... or do anything that sends an email. We'll talk more about that in a few minutes.