21.2.Functional Testing with Emails{Creating a Functional Test: bin/console make:functional-test}
Creating a Functional Test

Back to my real goal: writing a functional test for the registration page. Because a successful registration causes an email to be sent... I'm curious how that will work. Will an email actually be sent to Mailtrap? Do we want that?
To create the test, be lazy and run:

php bin/console make:functional-test
And... we immediately get an error: we're missing some packages. I'll copy the composer require browser-kit part. Panther isn't technically needed to write functional tests... and this error message is fixed in a newer version of this bundle. But, Panther is an awesome way to write functional tests that rely on JavaScript.