21.5.Functional Testing with Emails{Writing the Registration Functional Test; The test Environment Doesn't Read .env.local}
Writing the Registration Functional Test

We won't go too deep into the details of how to write functional tests, but it's a pretty simple idea. First, we create a $client object - which is almost like a "browser": it helps us make requests to our app. In this case, we want to make a GET request to /register to load the form.
The assertResponseIsSuccessful() method is a helper assertion from Symfony that will make sure the response wasn't an error or a redirect. Now... I'll remove the assertSelectorTextContains()... and paste in the rest of the test.
Let's see: this goes to /register, finds the Register button by its text, and then fills out all the form fields. These funny-looking values are literally the name attributes of each element if you looked at the source HTML. After submitting the form, we assert that the response is a redirect... which is an easy way to assert that the form submit was successful. If there's a validation error, it re-renders without redirecting.
We've used the registration form on this site... about 100 times. So we know it works... and so this test should pass. Whenever you say that something "should" work in programming... do you ever get the sinking feeling that you're about to eat your words? Ah, I'm sure nothing bad will happen in this case. Let's try it!
At your terminal, run just this test with:

php bin/phpunit tests/Controller/SecurityControllerTest.php
Deprecation notices of course... and... woh! It failed! And dumped some giant HTML... which is impossible to read... unless you go all the way to the top. Ah!

  Failed asserting that the Response is redirected: 500 internal server error.

And down in the HTML:

  Connection could not be established with host tcp://localhost:25

The test Environment Doesn't Read .env.local

Huh. That's coming from sending the email... but why is it trying to connect to localhost? Our config in .env.local is set up to talk to Mailtrap.
Well... there's a little gotcha about the .env system. I mean... it's a feature! When you're in the test environment, the .env.local file is not loaded. In every other situation - like the prod or the dev environments - it is loaded. But in test, it's not. It's madness!
Well, it definitely is surprising the first time you see this, but there is a good reason for it. In theory, your committed .env.test file should contain all the configuration needed for the test environment to work... on any machine. And so, you actually don't want your local values from .env.local to override the stuff in .env.test - that might break how the tests are supposed to behave.
The point is, since the .env.local file is not being loaded in our tests, it's using the .env settings for MAILER_DSN... which is connecting to localhost.