22.2.Email Delivery & Assertions in Tests{The Null Transport}
The Null Transport

Way earlier in the tutorial, I mentioned that the way an email is delivered is called a "transport". In .env, we're using the smtp transport to talk to the localhost server. In .env.local, this is also using the smtp transport to talk to the Mailtrap server. So far, smtp is the only transport we've seen.
Well, prepare to be amazed! Introducing the laziest, do-nothing... but mysteriously useful transport ever: the null transport! When you deliver an email via the null transport... your email goes... nowhere.
Hey! That's exactly what we want to do in the test environment! Inside .env.test, change MAILER_DSN to smtp://null.

Side note! This syntax changed in Symfony 4.4 to null://default - where the start of the string defines the transport type. We'll talk more about transports in a few minutes when we start using SendGrid.

 9 lines  .env.test
MAILER_DSN=smtp://null
# in Symfony 4.4 and higher, the syntax is
# MAILER_DSN=null://default
Anyways, let's try the test now:

php bin/phpunit tests/Controller/SecurityControllerTest.php
It passes and... yea! There were no email sent to Mailtrap. The test also ran about twice as fast.