22.3.Email Delivery & Assertions in Tests{Using the Null Transport by Default?}
Using the Null Transport by Default?

But wait, there's more! The null transport is perfect for the test environment. And... it might also be a good candidate as the default transport.
Hear me out. If a new developer cloned this project, they would not have a .env.local file. And so, out-of-the-box, mailer would use the smtp://localhost setting. What if this developer was really a designer that wanted to work on styling the registration process. Well... surprise! The moment they submit the form successfully, they'll be congratulated with a lovely 500 error. And they'll be off to find you to figure out how to fix it. That's no good for anyone.
That's why using the null transport in .env might be a perfect default. Then, if someone actually wants to test how the emails look, then they can take some time to configure their .env.local file to use Mailtrap.
Let's do this: change MAILER_DSN to smtp://null. Use null://default on Symfony 4.4 or higher.
Over in .env.test, we don't need to override anything. So, remove MAILER_DSN.