23.1.SendGrid & All About Transports{All About Transports;Let Symfony Guide You to Configure the Transport}
In .env, we're using the null transport. In .env.local, we're overriding that to send to Mailtrap. This is great for development, but it's time for our app to grow up, get a job, and join the real world. It's time for our app to send real emails through a real email server.
To do that, I recommend using a cloud-based email service... and Symfony Mailer can send to any service that supports the SMTP protocol... which is all of them. We did this for Mailtrap using the {username}:{password}@{server}:{port} syntax.
But to make life even nicer, Mailer has special support for the most common email services, like SendGrid, Postmark, Mailgun, Amazon SES and a few others. Let's use SendGrid.
Before we even create an account on SendGrid, we can jump in and start configuring it. In .env.local, comment-out the Mailtrap MAILER_DSN and replace it with MAILER_DSN=smtp://sendgrid. In Symfony 4.4, the syntax changed to sendgrid://default.

All About Transports

So far, we've seen two transports - two ways of delivering emails: the smtp transport and the null transport. Symfony also has a sendgrid transport, as well as a mailgun transport amazonses transport and many others.
In Symfony 4.3, you choose which transport you want by saying smtp:// and then the name of one of those transports, like null or sendgrid. In Symfony 4.4 and higher, this is different. The syntax is the transport name, like null or sendgrid then :// and whatever other options that transport needs. The word default is a dummy placeholder that's used when you don't need to configure a "server", like for the null transport or for sendgrid, because that transport already knows internally what the address is to the SendGrid servers.
Anyways, whether you're in Symfony 4.3 with the old syntax or Symfony 4.4 with the new one, this is how you say: "I want to deliver emails via the SendGrid transport".
At this point, some of you might be screaming
  Wait! That can't possibly be all the config we need to send emails!
And you're 1000% percent correct. This doesn't contain any SendGrid username, or API key. Heck, we haven't even created a SendGrid account yet! All true, all true. But let's... try it anyways. Because, Symfony is going to guide us through the process. How nice!

Let Symfony Guide You to Configure the Transport

Head over to the browser and refresh. Woh! An immediate error:

  Unable to send emails via Sendgrid as the bridge is not installed.
This is another example of Symfony making it easy to do something... but without bloating our project with stuff we don't need