23.2.SendGrid & All About Transports{Let Symfony Guide You to Configure the Transport: composer require symfony/sendgrid-mailer}
Now that we do want to use Sendgrid, it helps us install the required library. Copy the composer require line, spin over to your terminal and paste:

composer require symfony/sendgrid-mailer
Ooh, this package came with a recipe! Let's see what it did:

git status
In addition to the normal stuff, this also modified our .env file. Let's see how:

git diff .env
Cool! The recipe added a new section to the bottom! Back in our editor, let's see what's going on in .env:

 53 lines  .env
###> symfony/sendgrid-mailer ###
# SENDGRID_KEY=
# MAILER_DSN=smtp://$SENDGRID_KEY@sendgrid
###
Yea... this makes sense. We know that mailer is configured via a MAILER_DSN environment variable... and so when we installed the SendGrid mailer package, its recipe added a suggestion of how that variable should look in order to work with SendGrid.

SendGrid Symfony 4.4 Config Format
Now, two important notes about this. First, when you install this package in Symfony 4.4, the config added by the recipe will look a bit different: it will add just one line:

MAILER_DSN=sendgrid://KEY@default
Like we just talked about, this is because Symfony 4.4 changed the config format: the "transport type" is now at the beginning. The KEY is a placeholder: we'll replace with a real API key in a few minutes. And the @default part just tells the SendGrid transport to send the message to whatever the actual SendGrid hostname is.... we don't need to worry about configuring that.

A Note about Environment Variables inside Environment Variables
Now, if you look at the config that Symfony 4.3 uses, you'll notice the second important thing: this defines two environment variables. Gasp! It defines SENDGRID_KEY and then MAILER_DSN. This... is just a config trick. See how the MAILER_DSN value contains $SENDGRID_KEY? It's using that variable: it's environment variables inside environment variables! With this setup, you could commit this MAILER_DSN value to .env and then only need to override SENDGRID_KEY in .env.local.
This idea - the idea of using environment variables inside environment variables totally works in Symfony 4.4. But to keep the config a bit simpler, in Symfony 4.4 - you won't see this two-variable system in the recipe. Instead, we'll configure the entire MAILER_DSN value. After all, it's a pretty short string.
Next... let's actually do that configuration! It's time to create a SendGrid account and start using it.