26.1.Events & Overriding "Recipients"{ Hooking into Mailer: MessageEvent; Built-in Listener: EnvelopeListener; Setting up the Dev Email}
First, while we're developing, if we decide to use Mailtrap, great: all of our emails will go there. But if we decide that we want to use SendGrid to send real emails while developing... it's a little trickier. For example, whenever you register, you would need to use a real email address. Otherwise, the email would never make it to your inbox.
So here's idea number 1: what if, in the dev environment only, we globally override the "to" of every email and send to ourselves. So even if we registered as space_cadet@example.com - the email would actually be delivered to our real address: ryan@symfonycasts.com for me. That would be cool!
My second idea is similar: instead of manually setting the from() on every email object... what if we hook into mailer and set this globally. That's less duplication and more consistency.

Hooking into Mailer: MessageEvent
The way to accomplish both of these is by leveraging an event. Whenever an email is sent through Mailer, internally, it dispatches one event called MessageEvent. Mailer itself comes with a couple of classes that can "listen" to this event. The most interesting one is called EnvelopeListener.

Built-in Listener: EnvelopeListener
I'll hit Shift+Shift and look for EnvelopeListener so we can see inside. Start by looking for getSubscribedEvents(). Yep! This is listening on MessageEvent. Here's the idea: if you used this class, you could instantiate it and pass a custom sender or a custom array of recipients. Then, whenever an email is sent, the onMessage() method would be called and it would override that stuff on the email.
I love it! Even though this class lives inside Mailer, Symfony doesn't activate it by default: it's not currently being used. In Symfony 4.4, some new config options were been added so you can activate & configure it easily:

# config/packages/mailer.yaml
# or config/packages/dev/mailer.yaml for only the dev environment
framework:
  mailer:
    envelope:
      sender: 'sender@example.org'
      recipients: ['redirected@example.org']
But in Symfony 4.3, if we want to use this class, we need to activate it manually... which is kinda fun anyways.
So here's the plan: to start, in the development environment only, I want all emails to actually be sent to ryan@symfonycasts.com, regardless of the to() address on the email.

Setting up the Dev Email
To do this, in .env, let's create a brand new, shiny environment variable: DEV_MAIL_RECIPIENT set to, how about, someone@example.com.
That's not a real email, because each developer should need to copy this variable, open their own .env.local file, and customize it to whatever they want.