27.1.Setting "From" Globally{From Versus Sender}
I don't like to have this ->from() on every single email that I create. This will probably always be the same, so let's set it globally.
We know that Mailer dispatches an event each time it sends an email. So, we could probably create a listener for that event and set the from address from there!
But wait. A minute ago, we configured EnvelopeListener as a service in the dev environment and used it to globally override the recipients. This class also allows us to pass a "sender" as the first argument. If we did, it would override the sender on this "envelope" thing.
So, is setting the from globally as easy as passing a value to the first argument of EnvelopeListener? Is this video about 10 seconds from being over?

From Versus Sender
Sadly... no. Remember when I mentioned that an email is two parts: a message and then an envelope around that message? When you set the ->to() on an Email, that goes into the message. The recipients is what goes on the envelope... which totally impacts where the email is delivered, but does not impact who the email appears to be addressed to when reading the email.
The same is true when it comes to from() versus "sender". But this... is even more subtle. The "sender" is the address that's written on the envelope and the from is what actually goes into the message - this is the part that the user will see when reading the email. It's a weird distinction: it's like if someone mailed a letter on your behalf: they would be the sender - with their address on the envelope. But when you opened the envelope, the message inside would be signed from you.
The point is, setting the "sender" is not enough. When we set the from(), Mailer does automatically use that to set the "sender" on the envelope... unless it was set explicitly. But it does not do it the other way around: if we removed the ->from() line and only set the sender, Mailer would give us a huge error because the message would have no from.
So what does this all mean? It means EnvelopeListener can't help us: we need to override the "from", not the "sender". No problem: let's create our own event listener.